[chore] textStack 밑에 여백 추가
